chatbot/requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# OpenAI Agents SDK
openai-agents>=0.6.4
litellm>=1.0.0

# Databricks
databricks-connect>=13.0.0

# Production logging
structlog>=23.0.0

# CLI Interface
rich>=13.7.0

# Additional utilities
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.12.0
flake8>=6.1.0
